+++ src/backend/api/application.py
@@ -53,25 +53,8 @@ def create_app(config_name='dev'):
 
     app.register_blueprint(bp)
 
-
     from . import tasks
-    app.celery = make_celery(app)
+    app.celery = tasks.make_celery(app)
 
     return app
 
-
-def make_celery(app):
-    celery = Celery(
-        app.import_name,
-        backend=app.config['CELERY_RESULT_BACKEND'],
-        broker=app.config['CELERY_BROKER_URL']
-    )
-    celery.conf.update(app.config)
-
-    class ContextTask(celery.Task):
-        def __call__(self, *args, **kwargs):
-            with app.app_context():
-                return self.run(*args, **kwargs)
-
-    celery.Task = ContextTask
-    return celery
