+++ src/backend/api/tasks/__init__.py
@@ -1,1 +1,2 @@
-from .flask_celery import *
+from .celery_app import make_celery
+from .celery_tasks import *
